Add inputrc in the install

+++ install.sh
@@ -5,7 +5,7 @@ if [[ -d ~/.dotfiles  ]]; then
     if [[ -d ~/.dotfiles/vim ]]; then
 	mkdir -p ~/.vim/{view,undo,spell,pack/plugins/{start,opt}}
 	for i in $(ls ~/.dotfiles/vim/plugins); do
-	    ln -sf ~/.dotfiles/vim/plugins/"$i" ~/.vim/pack/plugins/start/"$i"
+	    [[ -f "$i" ]] && ln -sf ~/.dotfiles/vim/plugins/"$i" ~/.vim/pack/plugins/start/"$i"
 	done
 	[[ -f ~/.dotfiles/vim/vimrc ]] && ln -sf ~/.dotfiles/vim/vimrc  ~/.vim/vimrc
     fi
@@ -34,6 +34,12 @@ if [[ -d ~/.dotfiles  ]]; then
 	# if [[ -f ~/.dotfiles/bash/bashrc ]]; then
 	# fi
     fi
+
+    # readline
+    if [[ -d ~/.dotfiles/readline ]]; then
+	mkdir -p ~/.config/readline
+	[[ -f ~/.dotfiles/readline/inputrc ]] && ln -sf ~/.dotfiles/readline/inputrc  ~/.config/readline/inputrc
+    fi
 fi
 
 # TODO
@@ -41,6 +47,3 @@ fi
 # add .bashrc
 # add .tmuxrc
 # make a function to check if link / absent, etc. and make the link
-
-
-
